requirements.txt
fastapi
uvicorn
python-multipart
httpx
pillow
numpy
